+++ app/dashboards/room_assignment_dashboard.rb
@@ -2,8 +2,8 @@
 
 require 'administrate/base_dashboard'
 
-# administrate dashboard for rooms
-class RoomDashboard < Administrate::BaseDashboard
+# Administrate dashboard for RoomAssignments
+class RoomAssignmentDashboard < Administrate::BaseDashboard
   # ATTRIBUTE_TYPES
   # a hash that describes the type of each of the model's fields.
   #
@@ -11,13 +11,13 @@ class RoomDashboard < Administrate::BaseDashboard
   # which determines how the attribute is displayed
   # on pages throughout the dashboard.
   ATTRIBUTE_TYPES = {
-    suite: Field::BelongsTo,
-    users: Field::HasMany,
+    draw_membership: Field::BelongsTo,
+    room: Field::BelongsTo,
+    user: Field::HasOne,
+    group: Field::HasOne,
     id: Field::Number,
-    number: Field::String,
     created_at: Field::DateTime,
-    updated_at: Field::DateTime,
-    beds: Field::Number
+    updated_at: Field::DateTime
   }.freeze
 
   # COLLECTION_ATTRIBUTES
@@ -26,37 +26,36 @@ class RoomDashboard < Administrate::BaseDashboard
   # By default, it's limited to four items to reduce clutter on index pages.
   # Feel free to add, remove, or rearrange items.
   COLLECTION_ATTRIBUTES = %i(
-    number
-    suite
-    users
     id
+    user
+    room
+    group
   ).freeze
 
   # SHOW_PAGE_ATTRIBUTES
   # an array of attributes that will be displayed on the model's show page.
   SHOW_PAGE_ATTRIBUTES = %i(
-    suite
-    users
     id
-    number
+    user
+    room
+    draw_membership
+    group
     created_at
     updated_at
-    beds
   ).freeze
 
   # FORM_ATTRIBUTES
   # an array of attributes that will be displayed
   # on the model's form (`new` and `edit`) pages.
   FORM_ATTRIBUTES = %i(
-    suite
-    number
-    beds
+    draw_membership
+    room
   ).freeze
 
-  # Overwrite this method to customize how rooms are displayed
+  # Overwrite this method to customize how room assignments are displayed
   # across all pages of the admin dashboard.
   #
-  def display_resource(room)
-    "Room #{room.number}"
+  def display_resource(room_assignment)
+    "Room Assignment ##{room_assignment.id}"
   end
 end